messages v2
this version checks the users answer against the message text box, will need some error handling with capitals and spaces (maybe using split?), also this version prints the alphabets in the debug log. The debug log will also print the wrong or right (next steps will be to congratulate on screen and progress level or reset to retry. Next version aim is to show the encrypted message.